Stable ANN

+++ executor.cpp
@@ -6,8 +6,8 @@ void Executor::setANN(const char *fileName){
     _ann = fann_create_from_file(fileName);
 }
 
-void Executor::execute(QVector<qreal> input, QVector<qreal> &output){
+void Executor::execute(QVector<float> input, QVector<float> &output){
     fann_type *result = (fann_run(_ann, input.data()));
-    std::vector<double> stdOutput(result, result + sizeof(result)/sizeof(fann_type));
-    output = QVector<qreal>::fromStdVector(stdOutput);
+    std::vector<float> stdOutput(result, result + sizeof(result)/sizeof(fann_type));
+    output = QVector<float>::fromStdVector(stdOutput);
 }
